refactored area and capability generations. updated README to include a reference to agentic scripts.
