+++ src/area_generation/moderator.py
@@ -1,0 +1,254 @@
+"""Area moderator agent for managing the debate process."""
+
+import json
+import logging
+import traceback
+from pathlib import Path
+from typing import Dict, List
+
+from autogen_core import (
+    DefaultTopicId,
+    MessageContext,
+    RoutedAgent,
+    default_subscription,
+    message_handler,
+)
+from autogen_core.models import (
+    ChatCompletionClient,
+    SystemMessage,
+    UserMessage,
+)
+
+from ..utils.agentic_prompts import (
+    AREA_MODERATOR_MERGE_PROMPT,
+    AREA_MODERATOR_SYSTEM_MESSAGE,
+    AREA_FINALIZATION_INSTRUCTION,
+)
+from .messages import (
+    Domain,
+    AreaProposalRequest,
+    ScientistAreaProposal,
+    ScientistRevisionRequest,
+)
+
+log = logging.getLogger("agentic_area_gen.moderator")
+
+
+@default_subscription
+class AreaModerator(RoutedAgent):
+    """Moderator that merges scientist proposals and manages iteration."""
+
+    def __init__(
+        self,
+        model_client: ChatCompletionClient,
+        num_scientists: int,
+        num_final_areas: int,
+        max_round: int,
+        output_dir: Path,
+    ) -> None:
+        super().__init__("Area Moderator")
+        self._model_client = model_client
+        self._num_scientists = num_scientists
+        self._num_final_areas = num_final_areas
+        self._max_round = max_round
+        self._output_dir = output_dir
+        self._round = 0
+        self._proposals_buffer: Dict[int, List[ScientistAreaProposal]] = {}
+        self._domain = ""
+
+    @message_handler
+    async def handle_domain(self, message: Domain, ctx: MessageContext) -> None:
+        """Handle the domain message and initiate area proposal process."""
+        try:
+            log.info(f"Moderator received domain: {message.name}")
+            self._domain = message.name
+
+            # Send initial proposal request to scientists
+            await self.publish_message(
+                AreaProposalRequest(
+                    domain=message.name, num_areas=self._num_final_areas
+                ),
+                topic_id=DefaultTopicId(),
+            )
+
+        except Exception as e:
+            log.error(f"Error in Moderator handle_domain: {e}")
+            log.error(f"Traceback: {traceback.format_exc()}")
+            raise
+
+    @message_handler
+    async def handle_scientist_proposal(
+        self, message: ScientistAreaProposal, ctx: MessageContext
+    ) -> None:
+        """Handle area proposals from scientists."""
+        try:
+            log.info(
+                f"Moderator received proposal from Scientist {message.scientist_id} for round {message.round}"
+            )
+
+            self._proposals_buffer.setdefault(message.round, []).append(message)
+
+            if len(self._proposals_buffer[message.round]) == self._num_scientists:
+                log.info(
+                    f"Moderator received all proposals for round {message.round}, proceeding to merge"
+                )
+
+                # Get proposals from both scientists
+                proposals = self._proposals_buffer[message.round]
+                scientist_a_proposal = next(
+                    p.proposal for p in proposals if p.scientist_id == "A"
+                )
+                scientist_b_proposal = next(
+                    p.proposal for p in proposals if p.scientist_id == "B"
+                )
+
+                await self._merge_proposals(
+                    scientist_a_proposal, scientist_b_proposal, message.round
+                )
+
+        except Exception as e:
+            log.error(f"Error in Moderator handle_scientist_proposal: {e}")
+            log.error(f"Traceback: {traceback.format_exc()}")
+            raise
+
+    async def _merge_proposals(
+        self, scientist_a_proposal: str, scientist_b_proposal: str, round_num: int
+    ) -> None:
+        """Merge scientist proposals using LLM."""
+        try:
+            log.info(f"Moderator merging proposals for round {round_num}")
+
+            finalized_instruction = ""
+            finalized_field = ""
+            if round_num >= self._max_round - 1:
+                finalized_instruction = AREA_FINALIZATION_INSTRUCTION
+                finalized_field = ', "finalized": <true|false>'
+
+            prompt = AREA_MODERATOR_MERGE_PROMPT.format(
+                domain=self._domain,
+                scientist_a_proposal=scientist_a_proposal,
+                scientist_b_proposal=scientist_b_proposal,
+                num_final_areas=self._num_final_areas,
+                finalized_instruction=finalized_instruction,
+                finalized_field=finalized_field,
+            )
+
+            system_message = SystemMessage(content=AREA_MODERATOR_SYSTEM_MESSAGE)
+            user_message = UserMessage(content=prompt, source="user")
+
+            model_result = await self._model_client.create(
+                [system_message, user_message]
+            )
+
+            raw_content = model_result.content
+            if not isinstance(raw_content, str):
+                raw_content = str(raw_content)
+
+            # Check if finalized
+            is_finalized = False
+            try:
+                parsed = json.loads(raw_content)
+                is_finalized = parsed.get("finalized", False)
+            except Exception:
+                pass
+
+            if is_finalized or round_num >= self._max_round - 1:
+                log.info(f"Moderator finalizing areas after round {round_num}")
+                await self._finalize_areas(raw_content)
+            else:
+                log.info(
+                    f"Moderator sending merged proposal for revision in round {round_num}"
+                )
+                self._round = round_num + 1
+
+                # Send to scientists for revision
+                await self.publish_message(
+                    ScientistRevisionRequest(
+                        scientist_id="A",
+                        moderator_proposal=raw_content,
+                        round=self._round,
+                    ),
+                    topic_id=DefaultTopicId(),
+                )
+                await self.publish_message(
+                    ScientistRevisionRequest(
+                        scientist_id="B",
+                        moderator_proposal=raw_content,
+                        round=self._round,
+                    ),
+                    topic_id=DefaultTopicId(),
+                )
+
+        except Exception as e:
+            log.error(f"Error in Moderator _merge_proposals: {e}")
+            log.error(f"Traceback: {traceback.format_exc()}")
+            raise
+
+    async def _finalize_areas(self, final_areas: str) -> None:
+        """Save final areas to file."""
+        try:
+            log.info("Moderator finalizing and saving areas")
+
+            # Convert to the expected format with "areas" list
+            try:
+                json_start = final_areas.find("{")
+                if json_start != -1:
+                    json_part = final_areas[json_start:]
+                    parsed = json.loads(json_part)
+                else:
+                    parsed = json.loads(final_areas)
+
+                if "finalized" in parsed:
+                    del parsed["finalized"]  # Remove finalized flag
+
+                # Convert area_0, area_1 format to areas list
+                areas_list = []
+                i = 0
+                while f"area_{i}" in parsed:
+                    areas_list.append(parsed[f"area_{i}"])
+                    i += 1
+
+                final_format = {"areas": areas_list}
+                final_areas_json = json.dumps(final_format, indent=2)
+            except Exception as e:
+                log.warning(f"Could not parse final areas JSON: {e}")
+                final_areas_json = final_areas
+
+            self._save_areas_to_file(final_areas_json)
+            log.info("Area generation completed successfully")
+
+        except Exception as e:
+            log.error(f"Error in Moderator _finalize_areas: {e}")
+            log.error(f"Traceback: {traceback.format_exc()}")
+            raise
+
+    def _save_areas_to_file(self, areas: str) -> None:
+        """Save the generated areas to a file in the specified directory structure."""
+        try:
+            # Create the output directory if it doesn't exist
+            self._output_dir.mkdir(parents=True, exist_ok=True)
+            log.info(f"Created output directory: {self._output_dir}")
+
+            # Save as JSON file
+            areas_file = self._output_dir / "areas.json"
+
+            try:
+                # Try to parse as JSON first, if it's already JSON format
+                areas_data = json.loads(areas) if isinstance(areas, str) else areas
+            except json.JSONDecodeError as e:
+                log.warning(f"Areas string is not valid JSON, wrapping it: {e}")
+                # If not valid JSON, wrap in a simple structure
+                areas_data = {
+                    "raw_areas": areas,
+                    "error": "Original content was not valid JSON",
+                }
+
+            with open(areas_file, "w", encoding="utf-8") as f:
+                json.dump(areas_data, f, indent=2, ensure_ascii=False)
+
+            log.info(f"Areas saved to {areas_file}.")
+
+        except Exception as e:
+            log.error(f"Failed to save areas to file: {e}")
+            log.error(f"Traceback: {traceback.format_exc()}")
+            raise
