+++ src/capability_generation/generator.py
@@ -1,0 +1,159 @@
+"""Main capability generation orchestration functions."""
+
+import asyncio
+import json
+import logging
+import traceback
+from pathlib import Path
+
+from autogen_core import DefaultTopicId, SingleThreadedAgentRuntime
+from autogen_ext.models.openai import OpenAIChatCompletionClient
+from omegaconf import DictConfig
+
+from .messages import Area
+from .moderator import CapabilityModerator
+from .scientist import CapabilityScientist
+
+log = logging.getLogger("agentic_cap_gen.generator")
+
+
+async def generate_capabilities_for_area(
+    cfg: DictConfig, area: Area, output_dir: Path
+) -> None:
+    """Generate capabilities for a single area."""
+    try:
+        log.info(f"Generating capabilities for area: {area.name}")
+
+        runtime = SingleThreadedAgentRuntime()
+
+        await CapabilityScientist.register(
+            runtime,
+            "CapabilityScientistA",
+            lambda: CapabilityScientist(
+                model_client=OpenAIChatCompletionClient(
+                    model=cfg.agents.scientist_a.name,
+                    seed=cfg.agents.scientist_a.seed,
+                ),
+                scientist_id="A",
+                max_round=cfg.debate_cfg.max_round,
+                expected_capabilities=cfg.capabilities_cfg.num_capabilities_per_area,
+                domain=cfg.capabilities_cfg.domain,
+            ),
+        )
+
+        await CapabilityScientist.register(
+            runtime,
+            "CapabilityScientistB",
+            lambda: CapabilityScientist(
+                model_client=OpenAIChatCompletionClient(
+                    model=cfg.agents.scientist_b.name,
+                    seed=cfg.agents.scientist_b.seed,
+                ),
+                scientist_id="B",
+                max_round=cfg.debate_cfg.max_round,
+                expected_capabilities=cfg.capabilities_cfg.num_capabilities_per_area,
+                domain=cfg.capabilities_cfg.domain,
+            ),
+        )
+
+        await CapabilityModerator.register(
+            runtime,
+            "CapabilityModerator",
+            lambda: CapabilityModerator(
+                model_client=OpenAIChatCompletionClient(
+                    model=cfg.agents.moderator.name,
+                    seed=cfg.agents.moderator.seed,
+                ),
+                num_scientists=2,
+                num_capabilities=cfg.capabilities_cfg.num_capabilities_per_area,
+                max_round=cfg.debate_cfg.max_round,
+                output_dir=output_dir,
+                domain=cfg.capabilities_cfg.domain,
+            ),
+        )
+
+        # Start runtime and process the area
+        runtime.start()
+        await runtime.publish_message(area, DefaultTopicId())
+        log.info(f"Area message published: {area.name}")
+
+        # Wait for the runtime to stop when idle
+        try:
+            await runtime.stop_when_idle()
+            log.info(f"Completed generating capabilities for area: {area.name}")
+        except Exception as e:
+            log.error(f"Error while generating capabilities for area {area.name}: {e}")
+            raise
+
+    except Exception as e:
+        log.error(f"Error in generating capabilities for {area.name}: {e}")
+        log.error(f"Traceback: {traceback.format_exc()}")
+        raise
+
+
+async def generate_capabilities(cfg: DictConfig) -> None:
+    """Generate capabilities using multi-agent debate system for each area."""
+    try:
+        log.info("Starting capability generation process")
+
+        # Read areas from the areas.json file
+        areas_file = (
+            Path.home()
+            / cfg.debate_cfg.output_dir
+            / cfg.capabilities_cfg.domain
+            / cfg.exp_cfg.exp_id
+            / "areas"
+            / "areas.json"
+        )
+
+        if not areas_file.exists():
+            raise FileNotFoundError(f"Areas file not found: {areas_file}")
+
+        with open(areas_file, "r", encoding="utf-8") as f:
+            areas_data = json.load(f)
+
+        # Parse areas from the JSON data
+        areas = []
+        if isinstance(areas_data, dict) and "areas" in areas_data:
+            for area_dict in areas_data["areas"]:
+                if (
+                    isinstance(area_dict, dict)
+                    and "name" in area_dict
+                    and "description" in area_dict
+                ):
+                    areas.append(
+                        Area(
+                            name=area_dict["name"], description=area_dict["description"]
+                        )
+                    )
+
+        if not areas:
+            raise ValueError(f"No valid areas found in {areas_file}")
+
+        log.info(
+            f"Found {len(areas)} areas to process: {[area.name for area in areas]}"
+        )
+
+        output_dir = (
+            Path.home()
+            / cfg.debate_cfg.output_dir
+            / cfg.capabilities_cfg.domain
+            / cfg.exp_cfg.exp_id
+            / "capabilities"
+        )
+        log.info(f"Output directory: {output_dir}")
+
+        # Process each area individually with fresh agents
+        for i, area in enumerate(areas):
+            log.info(f"Processing area {i + 1}/{len(areas)}: {area.name}")
+
+            await generate_capabilities_for_area(cfg, area, output_dir)
+
+            log.info(f"Completed area {i + 1}/{len(areas)}: {area.name}")
+
+            await asyncio.sleep(1)
+
+    except Exception as e:
+        log.error(f"Error in generate_capabilities: {e}")
+        log.error(f"Traceback: {traceback.format_exc()}")
+        raise
